fix issue with Pellet
Unit test for conf 'examples/datatypes/stringtyped.conf' failed with an
internal error in Pellet reasoner. Not sure why this happens but apparently
it does not happen if we merge all axioms instead of only the logical
axioms. This commit seems to fix it but we should investigate
why this happens.